refactor(tests): Split high-priority narrative test files (#829)
Split two large test files exceeding 300-line limit into focused modules:

**narrativeStore.playerDecisionTracker.integration.test.ts (519 lines → 3 files)**
- userJourney.test.ts (255 lines) - Complete user journey integration tests
- robustness.test.ts (166 lines) - Integration robustness and error handling
- patterns.test.ts (93 lines) - Cross-session pattern building tests
- Enhanced testHelpers.ts with common mock data and utilities

**narrativeContextExtractor.test.ts (513 lines → 4 files)**
- location.test.ts (78 lines) - Location extraction tests
- situation.test.ts (101 lines) - Situation analysis tests
- characters.test.ts (97 lines) - Character presence detection tests
- edgeCases.test.ts (72 lines) - Edge cases and robustness tests
- Created testHelpers.ts (253 lines) with shared context builders

All files now comply with 300-line limit. Follows patterns from PR #825.

Related to #829
