--- src/state/__tests__/integration/narrativeStore.playerDecisionTracker.integration.test.ts
@@ -1,520 +1,0 @@
-/**
- * End-to-End Integration Tests for Issue #142
- * 
- * These tests verify the complete integration flow between narrativeStore.selectDecisionOption()
- * and PlayerDecisionTracker, including all supporting systems and edge cases.
- * 
- * Focus: Complete user journey and system integration
- */
-
-import { useNarrativeStore } from '../../narrativeStore';
-import { getTimestamp } from '@/lib/utils/timestamp';
-import { PlayerDecisionTracker } from '../../../lib/ai/playerDecisionTracker';
-import { DecisionOption } from '../../../types/narrative.types';
-import { ChoiceTypePreference } from '../../../types/personalization.types';
-
-// Create a test instance to avoid interfering with global state
-const createTestTracker = () => new PlayerDecisionTracker({
-  storageKey: 'test_integration_decisions',
-  maxDecisionsPerSession: 20,
-  maxTotalDecisions: 100
-});
-
-describe('NarrativeStore ↔ PlayerDecisionTracker End-to-End Integration (Issue #142)', () => {
-  let testTracker: PlayerDecisionTracker;
-
-  beforeEach(() => {
-    // Use fake timers for deterministic timestamp generation
-    jest.useFakeTimers();
-    jest.setSystemTime(new Date('2025-01-15T12:00:00Z'));
-
-    // Reset narrative store
-    useNarrativeStore.setState({
-      segments: {},
-      sessionSegments: {},
-      decisions: {},
-      sessionDecisions: {},
-      endedSessions: {},
-      currentEnding: null,
-      isGeneratingEnding: false,
-      endingError: null,
-      loading: false,
-      error: null
-    });
-
-    // Create fresh test tracker
-    testTracker = createTestTracker();
-    testTracker.clearDecisions();
-  });
-
-  afterEach(() => {
-    testTracker.clearDecisions();
-    jest.useRealTimers();
-  });
-
-  describe('Complete User Journey Integration', () => {
-    it('should track player choices throughout a complete game session', () => {
-      // INTEGRATION TEST: Complete flow from narrative segments through decision selection to pattern analysis
-      
-      const store = useNarrativeStore.getState();
-      const sessionId = 'integration-session-1';
-      const characterId = 'player-character';
-      const worldId = 'fantasy-world';
-
-      // Step 1: Establish narrative context with multiple segments
-      store.addSegment(sessionId, {
-        worldId,
-        content: 'You enter the bustling marketplace of Rivertown, filled with merchants and travelers.',
-        type: 'scene',
-        metadata: {
-          tags: ['marketplace', 'social', 'entry'],
-          location: 'Rivertown Marketplace',
-          mood: 'neutral'
-        },
-        updatedAt: getTimestamp(),
-        timestamp: new Date()
-      });
-
-      store.addSegment(sessionId, {
-        worldId,
-        content: 'A distressed merchant approaches you, wringing his hands nervously.',
-        type: 'dialogue',
-        characterIds: [characterId, 'merchant-desperate'],
-        metadata: {
-          tags: ['encounter', 'merchant', 'distress'],
-          location: 'Rivertown Marketplace'
-        },
-        updatedAt: getTimestamp(),
-        timestamp: new Date()
-      });
-
-      store.addSegment(sessionId, {
-        worldId,
-        content: '"Please, adventurer! Bandits stole my entire shipment on the road from Millhaven!"',
-        type: 'dialogue',
-        characterIds: [characterId, 'merchant-desperate'],
-        metadata: {
-          tags: ['plea', 'backstory', 'quest-hook']
-        },
-        updatedAt: getTimestamp(),
-        timestamp: new Date()
-      });
-
-      // Step 2: Create decision with contextually appropriate options
-      const decisionId = store.addDecision(sessionId, {
-        prompt: 'The merchant begs for your help recovering his stolen goods. What do you do?',
-        options: [
-          { 
-            id: 'help-free', 
-            text: 'Offer to help recover the goods for free, out of the goodness of your heart' 
-          },
-          { 
-            id: 'help-payment', 
-            text: 'Agree to help, but demand fair payment for the dangerous work' 
-          },
-          { 
-            id: 'negotiate-terms', 
-            text: 'Negotiate the terms carefully before committing to anything' 
-          },
-          { 
-            id: 'decline-politely', 
-            text: 'Politely decline and suggest he contact the town guard instead' 
-          },
-          { 
-            id: 'ignore-walk-away', 
-            text: 'Ignore his pleas and walk away to continue your own business' 
-          }
-        ] as DecisionOption[],
-        contextSummary: 'Merchant requests help with bandit problem',
-        decisionWeight: 'major'
-      });
-
-      // Step 3: Get the decision options before selection
-      let currentState = useNarrativeStore.getState();
-      const originalDecision = currentState.decisions[decisionId];
-      expect(originalDecision).toBeDefined();
-      const selectedOption = originalDecision.options.find(opt => opt.id === 'help-free');
-
-      // Simulate integration - player selects helpful option
-      // This should trigger the integration logic (when implemented)
-      store.selectDecisionOption(decisionId, 'help-free', characterId);
-
-      // Step 4: Get fresh state after selection
-      currentState = useNarrativeStore.getState();
-      const decision = currentState.decisions[decisionId];
-
-      // Simulate context extraction
-      const context = {
-        location: 'Rivertown Marketplace',
-        situation: 'merchant in distress',
-        charactersPresent: ['merchant-desperate']
-      };
-
-      // Simulate choice type inference
-      const choiceType: ChoiceTypePreference = 'helpful';
-
-      // Record decision in tracker (this would be automatic)
-      testTracker.recordDecision(
-        decision.prompt,
-        selectedOption!.text,
-        choiceType,
-        sessionId,
-        worldId,
-        context
-      );
-
-      // Step 5: Verify integration results
-      const trackedDecisions = testTracker.getSessionDecisions(sessionId);
-      expect(trackedDecisions).toHaveLength(1);
-
-      const trackedDecision = trackedDecisions[0];
-      expect(trackedDecision.prompt).toBe('The merchant begs for your help recovering his stolen goods. What do you do?');
-      expect(trackedDecision.choiceText).toBe('Offer to help recover the goods for free, out of the goodness of your heart');
-      expect(trackedDecision.choiceType).toBe('helpful');
-      expect(trackedDecision.sessionId).toBe(sessionId);
-      expect(trackedDecision.worldId).toBe(worldId);
-      expect(trackedDecision.context.location).toBe('Rivertown Marketplace');
-      expect(trackedDecision.context.situation).toBe('merchant in distress');
-      expect(trackedDecision.context.charactersPresent).toContain('merchant-desperate');
-
-      // Step 6: Continue the journey with more decisions
-      const followUpDecisionId = store.addDecision(sessionId, {
-        prompt: 'You find the bandits\' camp. They outnumber you 3 to 1. How do you approach?',
-        options: [
-          { id: 'sneak-stealth', text: 'Sneak in quietly and recover the goods without fighting' },
-          { id: 'negotiate-bandits', text: 'Approach openly and try to negotiate the return of the goods' },
-          { id: 'attack-direct', text: 'Charge in with weapons drawn and fight them head-on' },
-          { id: 'retreat-plan', text: 'Retreat and come back with town guards for backup' }
-        ] as DecisionOption[]
-      });
-
-      // Get the follow-up decision options before selection
-      currentState = useNarrativeStore.getState();
-      const originalFollowUpDecision = currentState.decisions[followUpDecisionId];
-      const followUpOption = originalFollowUpDecision.options.find(opt => opt.id === 'sneak-stealth');
-
-      // Player chooses stealth approach
-      store.selectDecisionOption(followUpDecisionId, 'sneak-stealth', characterId);
-
-      // Get fresh state for follow-up decision
-      currentState = useNarrativeStore.getState();
-      const followUpDecision = currentState.decisions[followUpDecisionId];
-      
-      testTracker.recordDecision(
-        followUpDecision.prompt,
-        followUpOption!.text,
-        'stealthy',
-        sessionId,
-        worldId,
-        {
-          location: 'Bandit Camp',
-          situation: 'tactical encounter',
-          charactersPresent: ['bandit-leader', 'bandit-1', 'bandit-2']
-        }
-      );
-
-      // Step 7: Analyze accumulated patterns
-      const patterns = testTracker.analyzeChoicePatterns();
-      
-      expect(patterns.dominantChoiceTypes).toContain('helpful');
-      expect(patterns.dominantChoiceTypes).toContain('stealthy');
-      expect(patterns.choiceDistribution.helpful).toBe(1);
-      expect(patterns.choiceDistribution.stealthy).toBe(1);
-      expect(patterns.patternStrength).toBeGreaterThan(0);
-
-      // Step 8: Verify narrative store integrity
-      const narrativeDecisions = store.getSessionDecisions(sessionId);
-      expect(narrativeDecisions).toHaveLength(2);
-      expect(narrativeDecisions[0].selectedOptionId).toBe('help-free');
-      expect(narrativeDecisions[1].selectedOptionId).toBe('sneak-stealth');
-    });
-
-    it('should handle complex multi-character decision scenarios', () => {
-      // INTEGRATION TEST: Multi-character decisions with rich context
-      
-      const store = useNarrativeStore.getState();
-      const sessionId = 'party-session';
-      const playerCharacterId = 'player-char';
-      const companionCharacterId = 'companion-char';
-      const worldId = 'adventure-world';
-
-      // Create party scenario context
-      store.addSegment(sessionId, {
-        worldId,
-        content: 'Your party stands before the ancient sealed door of the lost temple.',
-        type: 'scene',
-        characterIds: [playerCharacterId, companionCharacterId, 'temple-guardian'],
-        metadata: {
-          tags: ['temple', 'party', 'ancient', 'mystery'],
-          location: 'Temple of Forgotten Wisdom',
-          mood: 'mysterious'
-        },
-        updatedAt: getTimestamp(),
-        timestamp: new Date()
-      });
-
-      store.addSegment(sessionId, {
-        worldId,
-        content: 'The guardian spirit warns that opening the door will release both great knowledge and terrible danger.',
-        type: 'dialogue',
-        characterIds: [playerCharacterId, companionCharacterId, 'temple-guardian'],
-        metadata: {
-          tags: ['warning', 'choice', 'consequence']
-        },
-        updatedAt: getTimestamp(),
-        timestamp: new Date()
-      });
-
-      // Player decision
-      const playerDecisionId = store.addDecision(sessionId, {
-        prompt: 'As party leader, do you order the door to be opened despite the warning?',
-        options: [
-          { id: 'open-door', text: 'Open the door - the knowledge is worth the risk' },
-          { id: 'heed-warning', text: 'Heed the guardian\'s warning and leave the temple sealed' },
-          { id: 'consult-party', text: 'Consult with your party members before deciding' },
-          { id: 'study-inscriptions', text: 'Study the door\'s inscriptions to understand the danger better' }
-        ] as DecisionOption[]
-      });
-
-      // Player chooses to consult party (diplomatic approach)
-      store.selectDecisionOption(playerDecisionId, 'consult-party', playerCharacterId);
-
-      // Simulate tracking for player decision
-      testTracker.recordDecision(
-        'As party leader, do you order the door to be opened despite the warning?',
-        'Consult with your party members before deciding',
-        'diplomatic',
-        sessionId,
-        worldId,
-        {
-          location: 'Temple of Forgotten Wisdom',
-          situation: 'leadership decision',
-          charactersPresent: ['companion-char', 'temple-guardian']
-        }
-      );
-
-      // Companion decision (influenced by player's choice)
-      const companionDecisionId = store.addDecision(sessionId, {
-        prompt: 'Your companion, seeing your thoughtful leadership, offers their opinion. What do they suggest?',
-        options: [
-          { id: 'support-caution', text: 'Support leaving the temple sealed for everyone\'s safety' },
-          { id: 'support-knowledge', text: 'Support opening the door to gain the ancient knowledge' },
-          { id: 'suggest-preparation', text: 'Suggest preparing defenses before opening the door' }
-        ] as DecisionOption[]
-      });
-
-      store.selectDecisionOption(companionDecisionId, 'suggest-preparation', companionCharacterId);
-
-      // Simulate tracking for companion decision (would be separate character tracking)
-      testTracker.recordDecision(
-        'Your companion, seeing your thoughtful leadership, offers their opinion. What do they suggest?',
-        'Suggest preparing defenses before opening the door',
-        'helpful',
-        sessionId,
-        worldId,
-        {
-          location: 'Temple of Forgotten Wisdom',
-          situation: 'party consultation',
-          charactersPresent: [playerCharacterId, 'temple-guardian']
-        }
-      );
-
-      // Verify complex scenario tracking
-      const allDecisions = testTracker.getAllDecisions();
-      expect(allDecisions).toHaveLength(2);
-
-      const playerDecisions = testTracker.getSessionDecisions(sessionId);
-      expect(playerDecisions).toHaveLength(2);
-
-      const patterns = testTracker.analyzeChoicePatterns();
-      expect(patterns.choiceDistribution.diplomatic).toBe(1);
-      expect(patterns.choiceDistribution.helpful).toBe(1);
-    });
-  });
-
-  describe('Integration Robustness and Error Handling', () => {
-    it('should maintain data integrity during integration failures', () => {
-      // INTEGRATION TEST: System resilience when tracking fails
-      
-      const store = useNarrativeStore.getState();
-      const sessionId = 'robust-session';
-      const characterId = 'test-character';
-
-      // Create normal decision
-      const decisionId = store.addDecision(sessionId, {
-        prompt: 'Test decision for robustness',
-        options: [
-          { id: 'option-1', text: 'First option' },
-          { id: 'option-2', text: 'Second option' }
-        ] as DecisionOption[]
-      });
-
-      // Record selection in narrative store with deterministic timestamp
-      const expectedTime = new Date('2025-01-15T12:00:00Z');
-      store.selectDecisionOption(decisionId, 'option-1', characterId);
-
-      // Get fresh state after selection
-      const updatedState = useNarrativeStore.getState();
-      const decision = updatedState.decisions[decisionId];
-      expect(decision.selectedOptionId).toBe('option-1');
-      expect(decision.characterId).toBe(characterId);
-      expect(decision.selectedAt).toBeInstanceOf(Date);
-      expect(decision.selectedAt!.getTime()).toBe(expectedTime.getTime());
-
-      // Verify no errors are set
-      const currentState = useNarrativeStore.getState();
-      expect(currentState.error).toBeNull();
-
-      // Verify decision is retrievable
-      const sessionDecisions = store.getSessionDecisions(sessionId);
-      expect(sessionDecisions).toHaveLength(1);
-      expect(sessionDecisions[0].id).toBe(decisionId);
-    });
-
-    it('should handle missing or incomplete context gracefully', () => {
-      // INTEGRATION TEST: Partial context scenarios
-      
-      const store = useNarrativeStore.getState();
-      const sessionId = 'minimal-context-session';
-      const characterId = 'test-character';
-
-      // Create decision without rich context
-      const decisionId = store.addDecision(sessionId, {
-        prompt: 'A choice with minimal context',
-        options: [
-          { id: 'minimal-option', text: 'Make the choice anyway' }
-        ] as DecisionOption[]
-      });
-
-      // Make selection
-      store.selectDecisionOption(decisionId, 'minimal-option', characterId);
-
-      // Simulate tracking with minimal context
-      testTracker.recordDecision(
-        'A choice with minimal context',
-        'Make the choice anyway',
-        'neutral',
-        sessionId,
-        'unknown-world', // Fallback world ID
-        {} // Empty context
-      );
-
-      // Verify tracking works with minimal data
-      const decisions = testTracker.getSessionDecisions(sessionId);
-      expect(decisions).toHaveLength(1);
-      expect(decisions[0].context).toEqual({});
-      expect(decisions[0].worldId).toBe('unknown-world');
-    });
-
-    it('should handle rapid successive decisions without data loss', () => {
-      // INTEGRATION TEST: Performance under rapid decision making
-      
-      const store = useNarrativeStore.getState();
-      const sessionId = 'rapid-session';
-      const characterId = 'speed-character';
-
-      const decisions = [];
-      const numDecisions = 10;
-
-      // Create and select multiple decisions rapidly
-      for (let i = 0; i < numDecisions; i++) {
-        const decisionId = store.addDecision(sessionId, {
-          prompt: `Rapid decision ${i + 1}`,
-          options: [
-            { id: `option-${i}`, text: `Choice ${i + 1}` }
-          ] as DecisionOption[]
-        });
-
-        store.selectDecisionOption(decisionId, `option-${i}`, characterId);
-
-        // Simulate rapid tracking
-        testTracker.recordDecision(
-          `Rapid decision ${i + 1}`,
-          `Choice ${i + 1}`,
-          'neutral',
-          sessionId,
-          'speed-world',
-          { situation: `rapid choice ${i + 1}` }
-        );
-
-        decisions.push(decisionId);
-      }
-
-      // Verify all decisions were tracked
-      const trackedDecisions = testTracker.getSessionDecisions(sessionId);
-      expect(trackedDecisions).toHaveLength(numDecisions);
-
-      // Verify narrative store integrity
-      const narrativeDecisions = store.getSessionDecisions(sessionId);
-      expect(narrativeDecisions).toHaveLength(numDecisions);
-
-      // Verify each decision has correct data (sort numerically by prompt)
-      const sortedTrackedDecisions = trackedDecisions.sort((a, b) => {
-        const aNum = parseInt(a.prompt.match(/\d+/)?.[0] || '0');
-        const bNum = parseInt(b.prompt.match(/\d+/)?.[0] || '0');
-        return aNum - bNum;
-      });
-      sortedTrackedDecisions.forEach((decision, index) => {
-        expect(decision.prompt).toBe(`Rapid decision ${index + 1}`);
-        expect(decision.choiceText).toBe(`Choice ${index + 1}`);
-      });
-    });
-  });
-
-  describe('Cross-Session Pattern Building', () => {
-    it('should build consistent player patterns across multiple game sessions', () => {
-      // INTEGRATION TEST: Long-term pattern accumulation
-      
-      const store = useNarrativeStore.getState();
-      const characterId = 'consistent-character';
-      const worldId = 'pattern-world';
-
-      // Simulate multiple game sessions with consistent helpful choices
-      const sessions = ['session-1', 'session-2', 'session-3'];
-      
-      sessions.forEach((sessionId, sessionIndex) => {
-        // Create decision for each session
-        const decisionId = store.addDecision(sessionId, {
-          prompt: `Session ${sessionIndex + 1}: Someone needs help. What do you do?`,
-          options: [
-            { id: `help-${sessionIndex}`, text: `Help them in session ${sessionIndex + 1}` },
-            { id: `ignore-${sessionIndex}`, text: `Ignore them in session ${sessionIndex + 1}` }
-          ] as DecisionOption[]
-        });
-
-        // Player consistently chooses to help
-        store.selectDecisionOption(decisionId, `help-${sessionIndex}`, characterId);
-
-        // Track the decision
-        testTracker.recordDecision(
-          `Session ${sessionIndex + 1}: Someone needs help. What do you do?`,
-          `Help them in session ${sessionIndex + 1}`,
-          'helpful',
-          sessionId,
-          worldId,
-          {
-            situation: `session ${sessionIndex + 1} help scenario`
-          }
-        );
-      });
-
-      // Analyze cross-session patterns
-      const patterns = testTracker.analyzeChoicePatterns();
-      
-      expect(patterns.dominantChoiceTypes[0]).toBe('helpful');
-      expect(patterns.choiceDistribution.helpful).toBe(3);
-      expect(patterns.patternStrength).toBeGreaterThan(90); // Very strong pattern
-
-      // Verify each session has decisions
-      sessions.forEach(sessionId => {
-        const sessionDecisions = testTracker.getSessionDecisions(sessionId);
-        expect(sessionDecisions).toHaveLength(1);
-      });
-
-      // Test personalization readiness
-      const recentDecisions = testTracker.getRecentDecisions(30); // Last 30 days
-      expect(recentDecisions).toHaveLength(3);
-      expect(recentDecisions.every(d => d.choiceType === 'helpful')).toBe(true);
-    });
-  });
-});
